bug: [remove applicationlogger package; flatten into 1] (FF-3110)
