--- eppoclient/applicationlogger/interface.go
@@ -1,11 +1,0 @@
-package applicationlogger
-
-type Logger interface {
-	Debug(args ...interface{})
-	Info(args ...interface{})
-	Infof(template string, args ...interface{})
-	Warn(args ...interface{})
-	Warnf(template string, args ...interface{})
-	Error(args ...interface{})
-	Errorf(template string, args ...interface{})
-}
