--- eppoclient/applicationlogger/zap_logger.go
@@ -1,42 +1,0 @@
-package applicationlogger
-
-import (
-	"go.uber.org/zap"
-)
-
-// ZapLogger The default logger for the Eppo SDK
-type ZapLogger struct {
-	logger *zap.Logger
-}
-
-func NewZapLogger(logger *zap.Logger) *ZapLogger {
-	return &ZapLogger{logger: logger}
-}
-
-func (z *ZapLogger) Debug(args ...interface{}) {
-	z.logger.Sugar().Debug(args...)
-}
-
-func (z *ZapLogger) Info(args ...interface{}) {
-	z.logger.Sugar().Info(args...)
-}
-
-func (z *ZapLogger) Infof(template string, args ...interface{}) {
-	z.logger.Sugar().Infof(template, args...)
-}
-
-func (z *ZapLogger) Warn(args ...interface{}) {
-	z.logger.Sugar().Warn(args...)
-}
-
-func (z *ZapLogger) Warnf(template string, args ...interface{}) {
-	z.logger.Sugar().Warnf(template, args...)
-}
-
-func (z *ZapLogger) Error(args ...interface{}) {
-	z.logger.Sugar().Error(args...)
-}
-
-func (z *ZapLogger) Errorf(template string, args ...interface{}) {
-	z.logger.Sugar().Errorf(template, args...)
-}
