+++ eppoclient/scrubbing_logger_test.go
@@ -1,4 +1,4 @@
-package applicationlogger
+package eppoclient
 
 import (
 	"testing"
